refactor(controller): restructure into resource-based controllers for better organization
