+++ src/config/diContainer.ts
@@ -43,8 +43,12 @@ import { IChatService } from '../services/Interface/chat/IChatService';
 import ChatService from '../services/Implementation/chat/chatService';
 
 //CATEGORY
+import { ICategoryController } from '../controllers/Interface/category/ICategoryController';
+import CategoryController from '../controllers/Implementation/category/categoryController';
+import { ICategoryService } from '../services/Interface/category/ICategoryService';
+import CategoryService from '../services/Implementation/category/categoryService';
 import { ICategoryRepository } from '../repositories/Interface/ICategoryRepository';
-import { CategoryRepository } from '../repositories/Implementation/categoryRepository';
+import  CategoryRepository  from '../repositories/Implementation/categoryRepository';
 
 //PAYMENT
 import { IPaymentController } from '../controllers/Interface/user/IPaymentController';
@@ -64,19 +68,46 @@ import { IExpenseRepository } from '../repositories/Interface/IExpenseRepository
 import ExpenseRepository from '../repositories/Implementation/expenseRepository';
 
 //GROUP
+import { IGroupController } from '../controllers/Interface/group/IGroupController';
+import GroupController from '../controllers/Implementation/group/groupController';
+import { IGroupService } from '../services/Interface/group/IGroupService';
+import GroupService from '../services/Implementation/group/groupService';
 import { IGroupRepository } from '../repositories/Interface/IGroupRepository';
 import GroupRepository from '../repositories/Implementation/groupRepository';
 
 //SLOT
 import { ISlotRepository } from '../repositories/Interface/ISlotRepository';
 import SlotRepository from '../repositories/Implementation/slotRepository';
-import { IGoalRepository } from '../repositories/Interface/IGoalRepository';
-import GoalRepository from '../repositories/Implementation/goalRepository';
+
+//REVIEW
 import { IReviewRepository } from '../repositories/Interface/IReviewRepository';
 import ReviewRepository from '../repositories/Implementation/reviewRepository';
+
+//GOAL
+import { IGoalRepository } from '../repositories/Interface/IGoalRepository';
+import GoalRepository from '../repositories/Implementation/goalRepository';
+
+//NOTIFICATION
 import { INotificationRepository } from '../repositories/Interface/INotificationRepository';
 import NotificationRepository from '../repositories/Implementation/notificationRepository';
-
+import { IGoalController } from '../controllers/Interface/goal/IGoalController';
+import GoalController from '../controllers/Implementation/goal/goalController';
+import { IGoalService } from '../services/Interface/goal/IGoalService';
+import GoalService from '../services/Implementation/goal/goalService';
+import { ISlotController } from '../controllers/Interface/slot/ISlotController';
+import SlotController from '../controllers/Implementation/slot/slotController';
+import { ISlotService } from '../services/Interface/slot/ISlotService';
+import SlotService from '../services/Implementation/slot/slotService';
+import { IComplaintController } from '../controllers/Interface/complaint/IComplaintController';
+import ComplaintController from '../controllers/Implementation/complaint/complaintController';
+import { IComplaintService } from '../services/Interface/complaint/IComplaintService';
+import ComplaintService from '../services/Implementation/complaint/complaintService';
+import { IComplaintRepository } from '../repositories/Interface/IComplaintRepository';
+import ComplaintRepository from '../repositories/Implementation/complaintRepository';
+import { IReviewController } from '../controllers/Interface/review/IReviewController';
+import ReviewController from '../controllers/Implementation/review/reviewController';
+import { IReviewService } from '../services/Interface/review/IReviewService';
+import ReviewService from '../services/Implementation/review/reviewService';
 
 
 //user
@@ -105,6 +136,8 @@ container.register<IChatService>('IChatService',{useClass:ChatService})
 container.register<IChatRepository>('IChatRepository',{useClass:ChatRepository})
 
 //category
+container.register<ICategoryController>('ICategoryController',{useClass:CategoryController})
+container.register<ICategoryService>('ICategoryService',{useClass:CategoryService})
 container.register<ICategoryRepository>('ICategoryRepository',{useClass:CategoryRepository})
 
 //payment
@@ -117,18 +150,31 @@ container.register('StripeSecretKey', { useValue: STRIPEKEY });
 container.register<IExpenseController>('IExpenseController',{useClass:ExpenseController})
 container.register<IExpenseService>('IExpenseService',{useClass:ExpenseService})
 container.register<IExpenseRepository>('IExpenseRepository',{useClass:ExpenseRepository})
-
+ 
 //group
+container.register<IGroupController>('IGroupController',{useClass:GroupController})
+container.register<IGroupService>('IGroupService',{useClass:GroupService})
 container.register<IGroupRepository>('IGroupRepository',{useClass:GroupRepository})
 
-//slot
+//slot  
+container.register<ISlotController>('ISlotController',{useClass:SlotController})
+container.register<ISlotService>('ISlotService',{useClass:SlotService})
 container.register<ISlotRepository>('ISlotRepository',{useClass:SlotRepository})
 
 //goal
+container.register<IGoalController>('IGoalController',{useClass:GoalController})
+container.register<IGoalService>('IGoalService',{useClass:GoalService})
 container.register<IGoalRepository>('IGoalRepository',{useClass:GoalRepository})
 
 //review 
+container.register<IReviewController>('IReviewController',{useClass:ReviewController})
+container.register<IReviewService>('IReviewService',{useClass:ReviewService})
 container.register<IReviewRepository>('IReviewRepository',{useClass:ReviewRepository})
 
 //notification
 container.register<INotificationRepository>('INotificationRepository',{useClass:NotificationRepository})
+
+//report
+container.register<IComplaintController>('IComplaintController',{useClass:ComplaintController})
+container.register<IComplaintService>('IComplaintService',{useClass:ComplaintService})
+container.register<IComplaintRepository>('IComplaintRepository',{useClass:ComplaintRepository})
