+++ src/controllers/Interface/category/ICategoryController.ts
@@ -1,0 +1,5 @@
+import { Request, Response } from 'express'
+
+export interface ICategoryController {
+    getCategories(req: Request, res: Response): Promise<Response>;
+}
