+++ src/controllers/Interface/complaint/IComplaintController.ts
@@ -1,0 +1,6 @@
+import {Request,Response} from 'express'
+
+export interface IComplaintController{
+    reportAdvisor(req:Request, res:Response): Promise<Response> 
+
+}
