+++ src/controllers/Interface/review/IReviewController.ts
@@ -1,0 +1,6 @@
+import {Request,Response} from 'express'
+
+export interface IReviewController{
+    createReview(req: Request, res:Response): Promise<Response>
+
+}
