+++ src/controllers/Interface/slot/ISlotController.ts
@@ -1,0 +1,6 @@
+import {Request,Response} from 'express'
+
+export interface ISlotController{
+    bookslot(req: Request, res: Response): Promise<void>
+    fetchSlotsByUser(req:Request, res:Response): Promise<Response>
+}
