+++ src/repositories/Interface/IComplaintRepository.ts
@@ -1,0 +1,6 @@
+import { IReport } from "../../models/reportSchema";
+
+export interface IComplaintRepository{
+    createReport(data:IReport):Promise<IReport>
+
+}
