+++ src/routes/user/userRouter.ts
@@ -6,24 +6,35 @@ import { IChatController } from '../../controllers/Interface/chat/IChatControlle
 import { IPaymentController } from '../../controllers/Interface/user/IPaymentController';
 import { AuthMiddleware } from '../../middleware/authMiddleware';
 import { IExpenseController } from '../../controllers/Interface/user/IExpenseController';
+import { ICategoryController } from '../../controllers/Interface/category/ICategoryController';
+import { IGroupController } from '../../controllers/Interface/group/IGroupController';
+import { ISlotController } from '../../controllers/Interface/slot/ISlotController';
+import { IGoalController } from '../../controllers/Interface/goal/IGoalController';
+import { IComplaintController } from '../../controllers/Interface/complaint/IComplaintController';
+import { IReviewController } from '../../controllers/Interface/review/IReviewController';
 
 const userController = container.resolve<IUserController>('IUserController');
 const chatController = container.resolve<IChatController>('IChatController')
 const paymentController = container.resolve<IPaymentController>('IPaymentController')
 const expenseController = container.resolve<IExpenseController>('IExpenseController')
+const categoryController = container.resolve<ICategoryController>('ICategoryController')
+const groupController = container.resolve<IGroupController>('IGroupController')
+const slotController = container.resolve<ISlotController>('ISlotController')
+const goalController = container.resolve<IGoalController>('IGoalController')
+const complaintController = container.resolve<IComplaintController>('IComplaintController')
+const reviewController = container.resolve<IReviewController>('IReviewController')
 const router = Router()
-
 
 router.post('/upload',AuthMiddleware.authorizeUser,uploadProfile.single('profilePic'), (req, res) =>{userController.uploadProfileImage(req, res)})
 router.patch('/editProfile',AuthMiddleware.authorizeUser, (req, res) =>userController.updateUser(req, res))
 router.get('/getExpenses/:userId', AuthMiddleware.authorizeUser,(req,res)=>expenseController.getExpenses(req,res))
 router.post('/createExpense/:userId',AuthMiddleware.authorizeUser,(req,res)=>expenseController.createExpense(req,res))
-router.get('/getCategories',AuthMiddleware.authorizeUser,(req,res)=>userController.getCategories(req,res))
-router.post('/createGroup',AuthMiddleware.authorizeUser,(req,res)=>userController.createGroup(req,res))
-router.get('/getUserGroups/:userId',AuthMiddleware.authorizeUser,(req,res)=>userController.getUserGroups(req,res))
-router.post('/addMember/:groupId',AuthMiddleware.authorizeUser,(req,res)=>userController.addMember(req,res))
-router.post('/addExpenseInGroup/:groupId',AuthMiddleware.authorizeUser,(req,res)=>userController.addExpenseInGroup(req,res))
-router.patch('/bookslot',AuthMiddleware.authorizeUser,(req,res)=>userController.bookslot(req,res))
+router.get('/getCategories',AuthMiddleware.authorizeUser,(req,res)=>categoryController.getCategories(req,res))
+router.post('/createGroup',AuthMiddleware.authorizeUser,(req,res)=>groupController.createGroup(req,res))
+router.get('/getUserGroups/:userId',AuthMiddleware.authorizeUser,(req,res)=>groupController.getUserGroups(req,res))
+router.post('/addMember/:groupId',AuthMiddleware.authorizeUser,(req,res)=>groupController.addMember(req,res))
+router.post('/addExpenseInGroup/:groupId',AuthMiddleware.authorizeUser,(req,res)=>groupController.addExpenseInGroup(req,res))
+router.patch('/bookslot',AuthMiddleware.authorizeUser,(req,res)=>slotController.bookslot(req,res))
 router.post('/sendMessage',AuthMiddleware.authorizeUser,(req,res)=>chatController.sendMessage(req,res))
 router.get('/fetchMessages/:senderId/:receiverId',AuthMiddleware.authorizeUser,(req,res)=>chatController.fetchMessages(req,res))
 router.get('/fetchUsers/:id',AuthMiddleware.authorizeUser,(req,res)=>chatController.fetchUsers(req,res))
@@ -34,14 +45,14 @@ router.post('/createChat',AuthMiddleware.authorizeUser,(req,res)=>chatController
 router.post('/uploadChatFile',AuthMiddleware.authorizeUser,uploadChatFile.single('file'),(req,res)=>{chatController.uploadChatFile(req,res)})
 router.post('/paymentInitiate',AuthMiddleware.authorizeUser,(req,res)=>paymentController.initiatePayment(req,res))
 router.post('/confirmPayment',AuthMiddleware.authorizeUser,(req,res)=>paymentController.confirmPayment(req,res))
-router.post('/reportAdvisor/:slotId',AuthMiddleware.authorizeUser,(req,res)=>userController.reportAdvisor(req,res))
-router.get('/fetchSlotsByUser/:userId',AuthMiddleware.authorizeUser,(req,res)=>userController.fetchSlotsByUser(req,res))
-router.post('/createReview',AuthMiddleware.authorizeUser,(req,res)=>userController.createReview(req,res))
-router.post('/createGoals/:userId',AuthMiddleware.authorizeUser,(req,res)=>userController.createGoal(req,res))
-router.get('/getGoals/:userId',AuthMiddleware.authorizeUser,(req,res)=>userController.getGoalsById(req,res))
-router.patch('/updateGoal/:id',AuthMiddleware.authorizeUser,(req,res)=>userController.updateGoal(req,res))
-router.delete('/deleteGoal/:id',AuthMiddleware.authorizeUser,(req,res)=>userController.deleteGoal(req,res))
-router.patch('/updateGoalProgress/:id',AuthMiddleware.authorizeUser,(req,res)=>userController.updateGoalProgress(req,res))
+router.post('/reportAdvisor/:slotId',AuthMiddleware.authorizeUser,(req,res)=>complaintController.reportAdvisor(req,res))
+router.get('/fetchSlotsByUser/:userId',AuthMiddleware.authorizeUser,(req,res)=>slotController.fetchSlotsByUser(req,res))
+router.post('/createReview',AuthMiddleware.authorizeUser,(req,res)=>reviewController.createReview(req,res))
+router.post('/createGoals/:userId',AuthMiddleware.authorizeUser,(req,res)=>goalController.createGoal(req,res))
+router.get('/getGoals/:userId',AuthMiddleware.authorizeUser,(req,res)=>goalController.getGoalsById(req,res))
+router.patch('/updateGoal/:id',AuthMiddleware.authorizeUser,(req,res)=>goalController.updateGoal(req,res))
+router.delete('/deleteGoal/:id',AuthMiddleware.authorizeUser,(req,res)=>goalController.deleteGoal(req,res))
+router.patch('/updateGoalProgress/:id',AuthMiddleware.authorizeUser,(req,res)=>goalController.updateGoalProgress(req,res))
 router.get('/getDashboardData/:userId',AuthMiddleware.authorizeUser,(req,res)=>userController.getDashboardData(req,res))
 router.get('/exportExpense/:userId',AuthMiddleware.authorizeUser,(req,res)=>expenseController.exportExpense(req,res))
 router.get('/getNotification/:userId',AuthMiddleware.authorizeUser,(req, res) => chatController.getNotifications(req, res));
