+++ src/services/Interface/category/ICategoryService.ts
@@ -1,0 +1,3 @@
+export interface ICategoryService{
+    getCategories():Promise<any[]>
+}
