+++ src/services/Interface/complaint/IComplaintService.ts
@@ -1,0 +1,6 @@
+import { IReport } from "../../../models/reportSchema";
+
+export interface IComplaintService{
+    reportAdvisor(slotId:string,userId:string,advisorId:string,reason:string,customReason:string):Promise<IReport>
+
+}
