+++ src/services/Interface/review/IReviewService.ts
@@ -1,0 +1,6 @@
+import { IReview } from "../../../models/reviewSchema";
+
+export interface IReviewService{
+    createReview(advisorId: string, userId: string, rating: number, review: string): Promise<IReview>;
+
+}
